// src/app/shared/interfaces/movieInterface.ts
export interface movies{
    id:number;
    cnt:number;
    img:string;
    resumen:string;
    nombre:string; 
    gender:number[];
    layla:boolean;
}

export interface reser{
    peli: string
    diaD: number
    mesD: number
    anioD: number
    diaR: number
    mesR: number
    anioR: number
    poster: string
    pop: number
}